Move Asset System and Convert Util

+++ shared/src/main/java/blue/made/angleshared/AngleInfo.java
@@ -4,7 +4,7 @@ package blue.made.angleshared;
  * Created by Sumner Evans on 2016/12/14.
  */
 public class AngleInfo {
-    public static final short VERSION_MAJOR = -2;
+    public static final short VERSION_MAJOR = (short) -2;
     public static final short VERSION_MINOR = 0;
     public static final short VERSION_PATCH = 0;
 }
